requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
groq
pymongo
requests
edge-tts
